+++ service/video/v1/room/participant/api_op_client.go
@@ -5,6 +5,7 @@ import (
 	"github.com/RJPearson94/twilio-sdk-go/client"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/room/participant/published_track"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/room/participant/published_tracks"
+	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/room/participant/subscribe_rules"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/room/participant/subscribed_track"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/room/participant/subscribed_tracks"
 )
@@ -19,6 +20,7 @@ type Client struct {
 
 	PublishedTrack   func(string) *published_track.Client
 	PublishedTracks  *published_tracks.Client
+	SubscribeRules   func() *subscribe_rules.Client
 	SubscribedTrack  func(string) *subscribed_track.Client
 	SubscribedTracks *subscribed_tracks.Client
 }
@@ -48,6 +50,12 @@ func New(client *client.Client, properties ClientProperties) *Client {
 			ParticipantSid: properties.Sid,
 			RoomSid:        properties.RoomSid,
 		}),
+		SubscribeRules: func() *subscribe_rules.Client {
+			return subscribe_rules.New(client, subscribe_rules.ClientProperties{
+				ParticipantSid: properties.Sid,
+				RoomSid:        properties.RoomSid,
+			})
+		},
 		SubscribedTrack: func(subcribedTrackSid string) *subscribed_track.Client {
 			return subscribed_track.New(client, subscribed_track.ClientProperties{
 				ParticipantSid: properties.Sid,