+++ service/video/v1/room/participant/subscribe_rules/api_op_client.go
@@ -1,10 +1,10 @@
-// Package subscribed_tracks contains auto-generated files. DO NOT MODIFY
-package subscribed_tracks
+// Package subscribe_rules contains auto-generated files. DO NOT MODIFY
+package subscribe_rules
 
 import "github.com/RJPearson94/twilio-sdk-go/client"
 
-// Client for managing subscribed track resources
-// See https://www.twilio.com/docs/video/api/track-subscriptions#st-resource for more details
+// Client for managing a specific subscribe rules resource
+// See https://www.twilio.com/docs/video/api/track-subscriptions#sl-resource for more details
 type Client struct {
 	client *client.Client
 
@@ -12,13 +12,13 @@ type Client struct {
 	roomSid        string
 }
 
-// ClientProperties are the properties required to manage the subscribed tracks resources
+// ClientProperties are the properties required to manage the subscribe rules resources
 type ClientProperties struct {
 	ParticipantSid string
 	RoomSid        string
 }
 
-// New creates a new instance of the subscribed tracks client
+// New creates a new instance of the subscribe rules client
 func New(client *client.Client, properties ClientProperties) *Client {
 	return &Client{
 		client: client,